購物車金額型別改為 decimal，DTO/服務優化
將 Cart/CartItem 金額欄位由 int 改為 decimal，提升精度。新增 CartDto 與 CartItemDto，API 回傳時包含商品名稱、單價、總價等資訊。CartService 計算 FinalTotal 改用資料庫加總，避免同步問題。ProductService 商品 DTO 加入 Price。移除 CreateCartItemDto 的 Price 欄位，強制以資料庫價格為主。同步調整 .NET 及相關套件版本至 9.0。

+++ southernTravel/Model/Cart.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public int MemberId { get; set; }
-        public int FinalTotal { get; set; } = 0;
+        public decimal FinalTotal { get; set; } = 0;
         public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
 
         public ICollection<CartItem> CartItems { get; set; } = new List<CartItem>();
